+++ src/services/smart-fields-values-injector.js
@@ -9,42 +9,50 @@ function SmartFieldsValuesInjector(record, modelName, fieldsPerModel) {
 
   this.getFieldsForHighlightedSearch = () => fieldsForHighlightedSearch;
 
+  // NOTICE: Field reference format: `modelName.property`.
+  function getReferencedModelName(field) {
+    return field.reference.split('.')[0];
+  }
+
   // eslint-disable-next-line
   function setSmartFieldValue(record, field, modelName) {
-    function addFieldForHighlightIfCandidate(fieldToCheck) {
-      if (fieldToCheck.type === 'String') {
-        fieldsForHighlightedSearch.push(fieldToCheck.field);
-      }
+    if (field.value) {
+      logger.warn(`DEPRECATION WARNING: Smart Fields "value" method is deprecated. Please use "get" method in your collection ${modelName} instead.`);
     }
 
+    let value;
     try {
-      let value;
-      if (field.value) {
-        logger.warn(`DEPRECATION WARNING: Smart Fields "value" method is deprecated. Please use "get" method in your collection ${modelName} instead.`);
-      }
-
-      try {
-        value = field.get ? field.get(record) : field.value(record);
-      } catch (error) {
-        logger.error(`Cannot retrieve the ${field.field} value because of an internal error in the getter implementation: ${error}`);
-      }
-
-      if (!_.isNil(value)) {
-        if (_.isFunction(value.then)) {
-          return value
-            .then((result) => {
-              record[field.field] = result;
-              addFieldForHighlightIfCandidate(field);
-            })
-            .catch((error) => {
-              logger.warn(`Cannot set the field.field value because of an unexpected error: ${error}`);
-            });
-        }
-        record[field.field] = value;
-        return addFieldForHighlightIfCandidate(field);
-      }
+      value = field.get ? field.get(record) : field.value(record);
     } catch (error) {
-      logger.warn(`Cannot set the ${field.field} value because of an unexpected error: ${error}`);
+      logger.error(`Cannot retrieve the ${field.field} value because of an internal error in the getter implementation: ${error}`);
+    }
+
+    if (!_.isNil(value)) {
+      if (!_.isFunction(value.then)) {
+        value = Promise.resolve(value);
+      }
+      return value
+        .then(async (smartFieldValue) => {
+          // NOTICE: If the Smart Field is a Smart Relationship (ie references another record type),
+          //         we also need to inject the values of the referenced records Smart Fields.
+          if (smartFieldValue && smartFieldValue.dataValues && field.reference) {
+            const smartFieldsValuesInjector = new SmartFieldsValuesInjector(
+              smartFieldValue,
+              getReferencedModelName(field),
+              fieldsPerModel,
+            );
+            await smartFieldsValuesInjector.perform();
+          }
+          // NOTICE: Update the record with the the Smart Field value.
+          record[field.field] = smartFieldValue;
+          // NOTICE: String fields can be highlighted.
+          if (field.type === 'String') {
+            fieldsForHighlightedSearch.push(field.field);
+          }
+        })
+        .catch((error) => {
+          logger.warn(`Cannot set the ${field.field} value because of an unexpected error: ${error}`);
+        });
     }
   }
 
@@ -68,7 +76,7 @@ function SmartFieldsValuesInjector(record, modelName, fieldsPerModel) {
         }
       } else if (field.reference && !_.isArray(field.type)) {
         // NOTICE: Set Smart Fields values to "belongsTo" associated records.
-        const modelNameAssociation = field.reference.split('.')[0];
+        const modelNameAssociation = getReferencedModelName(field);
         const schemaAssociation = Schemas.schemas[modelNameAssociation];
 
         if (schemaAssociation && !_.isArray(field.type)) {